+++ src/operators/tensor/implementations/tensor_fp32x32.cairo
@@ -86,8 +86,8 @@ impl FP32x32Tensor of TensorTrait<FP32x32> {
     }
 
     fn argmax(
-        self: @Tensor<FP32x32>, axis: usize, keepdims: Option<bool>, select_last_index: Option<bool>
-    ) -> Tensor<usize> {
+        self: @Tensor<FP32x32>, axis: i32, keepdims: Option<bool>, select_last_index: Option<bool>
+    ) -> Tensor<i32> {
         math::argmax::argmax(self, axis, keepdims, select_last_index)
     }
