+++ src/operators/tensor/implementations/tensor_i32.cairo
@@ -87,8 +87,8 @@ impl I32Tensor of TensorTrait<i32> {
     }
 
     fn argmax(
-        self: @Tensor<i32>, axis: usize, keepdims: Option<bool>, select_last_index: Option<bool>
-    ) -> Tensor<usize> {
+        self: @Tensor<i32>, axis: i32, keepdims: Option<bool>, select_last_index: Option<bool>
+    ) -> Tensor<i32> {
         math::argmax::argmax(self, axis, keepdims, select_last_index)
     }
